requirements.txt
chainlit
openai
pyttsx3
speechrecognition
pyaudio
python-dotenv
pydub
